+++ npcroomui/index.html
@@ -2,10 +2,9 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/NPC_wojak_meme.png" />
+    <link rel="icon" href="/NPC_wojak_meme.png" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="rgb(38, 45, 45)" />
-    <style rel="stylesheet" href="%PUBLIC_URL%/index.css"></style>
 
     <!-- Material UI Font -->
     <link rel="preconnect" href="https://fonts.googleapis.com">
@@ -14,14 +13,13 @@
 
     <meta
       name="description"
-      content="Web site created using create-react-app"
+      content="Web site created using Vite"
     />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
     <!--
       manifest.json provides metadata used when your web app is installed on a
       user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
     -->
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+    <link rel="manifest" href="/manifest.json" />
     <!--
       Notice the use of %PUBLIC_URL% in the tags above.
       It will be replaced with the URL of the `public` folder during the build.
@@ -36,6 +34,7 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
+    <script type="module" src="src/index.tsx"></script>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
