+++ npcroomui/src/index.tsx
@@ -1,12 +1,9 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
+import { createRoot } from 'react-dom/client';
 import App from './App';
 import './index.css';
 import { ThemeProvider, createTheme } from '@mui/material';
 
-//Create a custom theme:
 const theme = createTheme({
-
   palette:{
     primary:{
       main: '#1F2833',
@@ -14,8 +11,6 @@ const theme = createTheme({
     secondary:{
       main: '#66FCF1',
     }
-    
-
   },
   typography: {
     // Define typography options here
@@ -23,7 +18,6 @@ const theme = createTheme({
       'Grandstander',
       'sans-serif'
     ].join(','),
-
     allVariants: {
     
     },
@@ -31,14 +25,8 @@ const theme = createTheme({
 })
 
 
-
-const root = ReactDOM.createRoot(document.getElementById('root'));
-root.render(
-
-  <React.StrictMode> 
-    <ThemeProvider theme={theme}>
-      <App />
-    </ThemeProvider>
-    
-  </React.StrictMode>
-);
+const container = document.getElementById('root');
+if (container) {
+  const root = createRoot(container);
+  root.render(<ThemeProvider theme={theme}><App /></ThemeProvider>);
+}
